+++ lib/cli/commands/clean.js
@@ -1,6 +1,19 @@
 const { readdirSync, removeSync } = require('fs-extra')
-const keeps = require('./keep')
-const getConfig = require('./config')
+const getConfig = require('../config')
+
+const KEEPS = [
+  'themes',
+  'scripts',
+  'config.yml',
+  'CNAME',
+  'README.md',
+  'LICENSE',
+  '.gitignore',
+  'package.json',
+  'package-lock.json',
+  'node_modules',
+  '.git',
+]
 
 module.exports = {
   name: 'clean',
@@ -10,7 +23,7 @@ module.exports = {
     if (getConfig()) {
       try {
         readdirSync(process.cwd())
-          .filter(file => file[0] !== '.' && keeps.indexOf(file) === -1)
+          .filter(file => file[0] !== '.' && KEEPS.indexOf(file) === -1)
           .forEach(file => removeSync(file))
       } catch (e) {
         logger.error(e)
