+++ lib/cli/commands/generate.js
@@ -2,7 +2,7 @@ module.exports = {
   name: 'generate',
   fullName: 'generate',
   description: 'Generate static files',
-  action(args) {
+  action(args, logger) {
     console.log(args)
   },
 }
