+++ lib/cli/commands/init.js
@@ -1,15 +1,20 @@
 const { resolve, join } = require('path')
 const { copySync, outputFileSync, existsSync } = require('fs-extra')
-const ignores = require('./ignore')
+
+const IGNORES = [
+  'Thumbs.db',
+  '.DS_Store',
+  '*.swp',
+  'themes/',
+]
 
 module.exports = {
   name: 'init',
   fullName: 'init [folder]',
   description: 'Create new website',
   action({ _ }, logger) {
-    const dir = __dirname
     const cwd = process.cwd()
-    const assetsPath = resolve(dir, '../../assets')
+    const assetsPath = resolve(__dirname, '../../../assets')
     const [, path = ''] = _
     const ignoresPath = join(cwd, path, '.gitignore')
 
@@ -24,7 +29,7 @@ module.exports = {
     try {
       copySync(assetsPath, join(cwd, path), { filter })
       if (!existsSync(ignoresPath)) {
-        outputFileSync(ignoresPath, ignores)
+        outputFileSync(ignoresPath, IGNORES.join('\n'))
       }
       logger.success('Configure `config.yml` to start')
     } catch (e) {
