+++ lib/cli/commands/version.js
@@ -1,4 +1,4 @@
-const { version } = require('../../package.json')
+const { version } = require('../../../package.json')
 
 module.exports = {
   name: '--version',
