+++ lib/cli/config.js
@@ -4,11 +4,10 @@ const { join } = require('path')
 
 const renderer = new Renderer()
 const logger = new Logger()
-const cwd = process.cwd()
 
 module.exports = () => {
   try {
-    return renderer.renderFile('yaml', join(cwd, 'config.yml'))
+    return renderer.renderFile('yaml', join(process.cwd(), 'config.yml'))
   } catch (e) {
     logger.error('Cannot find `config.yml` or configuration error')
     return null
