--- lib/cli/ignore.js
@@ -1,8 +1,0 @@
-const ignores = [
-  'Thumbs.db',
-  '.DS_Store',
-  '*.swp',
-  'theme/',
-]
-
-module.exports = ignores.join('\n')
