+++ lib/cli/index.js
@@ -1,8 +1,8 @@
 const Logger = require('@acyort/logger')
-const version = require('./version')
-const clean = require('./clean')
-const init = require('./init')
-const generate = require('./generate')
+const version = require('./commands/version')
+const clean = require('./commands/clean')
+const init = require('./commands/init')
+const generate = require('./commands/generate')
 
 class C {
   constructor() {
