--- lib/cli/keep.js
@@ -1,13 +1,0 @@
-module.exports = [
-  'themes',
-  'scripts',
-  'config.yml',
-  'CNAME',
-  'README.md',
-  'LICENSE',
-  '.gitignore',
-  'package.json',
-  'package-lock.json',
-  'node_modules',
-  '.git',
-]
